pyproject.toml
[tool.poetry]
name = "putjanitor"
version = "0.1.0"
description = ""
authors = ["Lukas Martini <hello@lutoma.org>"]

[tool.poetry.dependencies]
python = "^3.8"
"putio.py" = "^8.6.0"
PyYAML = "^5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
